requirements.txt
distlib==0.3.8
filelock==3.15.4
platformdirs==4.2.2
virtualenv==20.26.3
asgiref==3.8.1
django==5.0.7
sqlparse==0.5.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
coreapi==2.3.3
setuptools==71.1.0
